Fixed checkstyle and moved DocumentAnalyserMultiplexer
