+++ apache-rat-core/src/test/java/org/apache/rat/document/impl/DocumentAnalyserMultiplexerTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  */ 
-package org.apache.rat.document.impl.util;
+package org.apache.rat.document.impl;
 
 import org.apache.rat.document.IDocumentAnalyser;
 import org.apache.rat.testhelpers.TestingDocument;
